+++ src/Auth0PatternServiceProvider.php
@@ -5,15 +5,15 @@ namespace FaithFM\Auth0Pattern;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Gate;
 use App\Repositories\AuthPermissionList;
+use FaithFM\Auth0Pattern\Http\Middleware\PatchedAuthenticationMiddleware;
+use Illuminate\Routing\Router;
 
 class Auth0PatternServiceProvider extends ServiceProvider
 {
     /**
      * Register any package services.
-     *
-     * @return void
      */
-    public function register()
+    public function register() :void
     {
         $this->app->bind(
             \Auth0\Login\Contract\Auth0UserRepository::class,
@@ -23,10 +23,8 @@ class Auth0PatternServiceProvider extends ServiceProvider
 
     /**
      * Bootstrap any package services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(Router $router): void
     {
         // (ONCE-OFF ON PACKAGE INSTALLATION)
         // Publish everything contained in the "templates" folder
@@ -42,7 +40,6 @@ class Auth0PatternServiceProvider extends ServiceProvider
             __DIR__.'/../clone/' => base_path(),
         ], 'auth-every-update-force-clones');
 
-
         // Create gates for each permission an application defines in the AuthPermissionList
         if (class_exists(AuthPermissionList::class)) {
             foreach (AuthPermissionList::DEFINED_PERMISSIONS as $permission) {
@@ -51,6 +48,9 @@ class Auth0PatternServiceProvider extends ServiceProvider
                 });
             }
         }
+
+        // Register our Authentication Middleware
+        $router->aliasMiddleware('auth.patched', PatchedAuthenticationMiddleware::class);
 
         // Create routes
         $this->loadRoutesFrom(__DIR__.'/routes/auth0pattern-web.php');