+++ src/Http/Middleware/StartSession.php
@@ -17,9 +17,8 @@ class StartSession extends BaseStartSession
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \Closure  $next
-     * @return mixed
      */
-    public function handle($request, Closure $next)
+    public function handle($request, Closure $next): mixed
     {
         if(\Request::has('api_token'))
         {